requirements.txt
asgiref==3.7.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Limiter==3.5.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
GitHub-Flask==3.2.0
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
importlib_resources==6.1.3
itsdangerous==2.1.2
Jinja2==3.1.3
limits==3.10.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.2
psycopg2==2.9.9
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
rich==13.7.1
SQLAlchemy==2.0.28
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
